agents/unix: avoid timeouts when getting iptables table list
Misconfigured DNS can cause timeouts when processing reverse queries,
which are used when iptables wants to display domain names instead of
IP addresses.

When obtaining the list of available iptables tables, the timeouts
stack up for each table name that needs to be checked, to the point
where the TA times out and is considered dead.

Since the output is redirected straight to /dev/null and will not be
seen or processed in any way, disable DNS lookups when obtaining the
table list.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>